Move generation and more
Improved the move generator:bishop, queen, rook and king moving now works(except castling), and pawns can do everything except en peasent. Knights is still not in there.
Added so you can see where you can place pieces, however you can place them however you want, because the moves are not restricted to semi legal moves currently.

Made it so the colorToMove changes when a move is made.

QOL:changed things that took byte to take int, it was a lot of casting before.

+++ Assets/Scripts/Moving/Move.cs
@@ -12,7 +12,7 @@ public struct Move
 		get {  return theMove; }
 	}
 
-	public Move(byte startSquare, byte targetSquare)
+	public Move(int startSquare, int targetSquare)
 	{
 		theMove = 0;
 		//0b 000000        000000         0000
